Extract TranscriptionEngine protocol + first app-target test using a module Mocks library
Carves the existing actor into a protocol (TranscriptionEngine) + concrete
conformer (DefaultTranscriptionEngine), adds a TranscriptionKitMocks
library shipping MockTranscriptionEngine, and migrates TranscriptionManager
to take any TranscriptionEngine via init.

The new VivaDictaTests/TranscriptionManagerTests.swift is the first
app-target test class that imports a module Mocks library
(TranscriptionKitMocks) - closes the gap where module mocks existed but
were never reachable from app-target tests.

Changes:

- Modules/TranscriptionKit:
  - Renames the existing actor TranscriptionEngine -> DefaultTranscriptionEngine
    (file renamed too).
  - Adds new protocol TranscriptionEngine (transcribe, preloadWhisperKitModel)
    in a fresh TranscriptionEngine.swift.
  - DefaultTranscriptionEngine: TranscriptionEngine via empty extension.
  - Adds TranscriptionKitMocks library target shipping
    MockTranscriptionEngine (Bev-style stubs/callCounts/captures).
- Modules/TranscriptionKit/Tests: 5 instantiation sites updated to
  DefaultTranscriptionEngine().
- Modules/{Cloud,Local}TranscriptionMocks: doc comments updated to
  reference DefaultTranscriptionEngine.ServiceFactory (the typealias is
  concrete-only).
- VivaDicta/Services/Transcription/TranscriptionManager.swift:
  private let engine: any TranscriptionEngine + init(engine:) with a
  DefaultTranscriptionEngine() default so existing call sites still work.
- VivaDicta.xcodeproj/project.pbxproj: 4 surgical edits to link
  TranscriptionKitMocks into the VivaDictaTests target (PBXBuildFile,
  Frameworks build phase, packageProductDependencies,
  XCSwiftPackageProductDependency).
- VivaDictaTests/TranscriptionManagerTests.swift: 4 proof-of-pattern
  tests injecting MockTranscriptionEngine through the protocol. Coverage
  is intentionally narrow (preload skip branches, error-path verification,
  setCurrentMode) - the manager's reads of UserDefaultsStorage.shared /
  AppGroupCoordinator.shared / CustomTranscriptionModelManager.shared
  keep the success path out of scope until those are DI'd too.

Test plan:
- swift build --package-path Modules/TranscriptionKit: ok (Default + Mocks
  + Tests target compile cleanly).
- xcodebuild ... build-for-testing: ok.
- xcodebuild ... test -only-testing:VivaDictaTests/TranscriptionManagerTests:
  4/4 passing.
- xcodebuild ... test (full suite): 428 tests pass, 0 failures.
